Greatly simplified send_console string split by providing direct nchar splits on a specified integer and ignoring newlines.
I intend to add an issue which will add an option on how to handle the splits through either a newline counting approach or direct nchar split. I will need to add a split count variable as well.

I added tests to account for multiple response status codes for the >nchar limit.